# api/swagger.yaml
openapi: 3.0.3
info:
  title: API REST Simple
  description: |-

    ## Importante:

    ### Antes de realizar cualquier accion, primero realizar una peticion POST a la **ruta 'auth/login'** mas abajo

    #### Datos necesarios dentro del cuerpo de la peticion (ejemplo):

    ```json
    {
      "nombre":"usuariotest@test.com"
      "password":"simpletest"
    }
    ```

    ### Una vez obtenido el token en la respuesta. introducirlo clickeando en el boton **'Authorize'**

    #### Disfruta las consultas 



    link de google:
    [google](https://www.google.cl)
  contact:
    email: felipe.daniel.henriquez@gmail.com
  version: 3.0.1
...
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /categorias/{id}:
  # Categories -  Get One 
    get:
      tags:
        - Categorias
      summary: Muestra la categoria con el id pasado por parametro
      description: solo requiere token. Accesible para cualquier tipo de usuario
      operationId: obtenerCategoria
      parameters:
        - name: id
          in: path
          description: 'id de la categoria'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '404':
          description: categoria no encontrada
  # Categories - Update
    put:
      tags:
        - Categorias
      summary: Actualiza una categoria
      description: Actualiza una categoria con el id (mongo objectId) pasado por la ruta
      operationId: actualizarCategoria
      parameters:
        - name: id
          in: path
          description: id de la categoria a actualizar
          required: true
          schema:
            type: string
      requestBody:
        description: Actualiza una categoria
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaUpdate'
      responses:
        default: 
          description: operacion exitosa
      security:
        - api_key: []
  # Categories - Delete
    delete:
      tags:
        - Categorias
      summary: Elimina una categoria
      description: Elimina la categoria con el id pasado por parametro
      operationId: eliminarCategoria
      parameters:
        - name: id
          in: path
          description: id de la categoria que se va a eliminar
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Categoria eliminada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria/'
      security:
        - api_key: []
...
components:
...
  schemas:
...
    Categoria:
      type: object
      properties:
        nombre:
          type: string
        estado:
          type: boolean
        usuario:
          type: string
      required:
        - nombre
        - estado
        - usuario
...
    CategoriaUpdate:
      type: object
      properties:
        nombre: 
          type: string
    # Schema productos
...
  securitySchemes:
    api_key:
      type: apiKey
      name: x-token
      in: header